+++ app/Services/MaintainableNotificationSchedulingService.php
@@ -43,7 +43,9 @@ class MaintainableNotificationSchedulingService
     public function updateScheduleOfMaintainable(Maintainable $maintainable)
     {
         // 1. il faut rechercher toutes les  scheduled_notifications avec le notification_type et le user_id ET l'asset_type
-        // dump('--- UPDATE SCHEDULE OF MAINTAINABLE ---');
+        dump('--- UPDATE SCHEDULE OF MAINTAINABLE ---');
+
+        dump($maintainable->maintainable);
 
         // $maintainable->refresh();
         $users = User::role('Admin')->get();
@@ -238,13 +240,23 @@ class MaintainableNotificationSchedulingService
 
     public function removeNotificationsForOldMaintenanceManager(Maintainable $maintainable, User $user)
     {
-        // dump('--- removeNotificationsForOldMaintenanceManager ---');
+        // only remove notification if the user has the maintenance manager role
+        if ($user->hasAnyRole('Admin'))
+            return;
+
+
         $notifications = $maintainable->maintainable->notifications()->where('user_id', $user->id)->get();
         // dump($notifications);
 
         if (count($notifications) > 0)
             foreach ($notifications as $notification) {
                 $notification->delete();
+            }
+
+        $interventions = $maintainable->maintainable->interventions;
+        if (count($interventions) > 0)
+            foreach ($interventions as $intervention) {
+                app(InterventionNotificationSchedulingService::class)->removeNotificationsForOldMaintenanceManager($intervention, $user);
             }
     }
 }
